+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Bug fixes
 
+* Fixed a crash when calling Table.toString() in debugger (#2429).
 * Revised `RealmResults.isLoaded()` description (#2895).
 
 ## 1.0.0